requirements.txt
azure-identity
azure-keyvault-secrets
chromadb>=0.4.4
openai
tqdm